bluesky-parent/bluesky-web/bluesky-web-gate/src/main/java/net/luversof/web/gate/board/service/BoardArticleService.java
package net.luversof.web.gate.board.service;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import io.github.luversof.boot.exception.BlueskyException;
import net.luversof.web.gate.board.client.BoardArticleClient;
import net.luversof.web.gate.board.client.BoardClient;
import net.luversof.web.gate.board.domain.BoardArticle;

@Service
public class BoardArticleService {
	
	@Autowired
	private BoardClient boardClient;
	
	@Autowired
	private BoardArticleClient boardArticleClient;
	
	public Page<BoardArticle> findByBoardAlias(String boardAlias, Pageable pageable) {
		
		var board = boardClient.findByAlias(boardAlias);
		if (board == null) {
			throw new BlueskyException("board.NOT_EXIST_BOARD");
		}
		
		return boardArticleClient.findByBoardAlias(boardAlias, pageable);
	}

}
